make width works again
support internal

+++ include/fast_io_core_impl/utils.h
@@ -481,7 +481,7 @@ inline constexpr output_iter my_copy(input_iter first,input_iter second,output_i
 	std::is_trivially_copyable_v<output_value_type>&&
 	(std::same_as<input_value_type,output_value_type>||(std::integral<input_value_type>&&std::integral<output_value_type>&&sizeof(std::is_trivially_copyable_v<input_value_type>)==sizeof(std::is_trivially_copyable_v<output_value_type>))))
 	{
-		my_copy_n(first,second-first,result);
+		my_copy_n(first,static_cast<std::size_t>(second-first),result);
 		return result+(second-first);
 	}
 	else
